# java-angular/book-social-network/book-network-api/src/main/java/koula/diallo/book/manage/UserResponse.java
package koula.diallo.book.manage;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class UserResponse {
    private String firstname;
    private String lastname;
    private boolean accountLocked;
    private boolean enable;
}